==> 1.DataCollection/2.SuicideRateIndcator_Collection.ipynb <==
<center>

# **Gather indicator data**

</center>

Some indicatores where from PAHO sources </br>
https://opendata.paho.org/en/core-indicators/download-dataset</br>


Others from WHO API</br>
https://www.who.int/data/gho/info/gho-odata-api


import pandas as pd
from prettytable import PrettyTable
import matplotlib.pyplot as plt
from collections import Counter
import requests
import io
import os

#Mount your Google Drive to Colab
from google.colab import drive
drive.mount('/content/drive')

os.listdir('/content/drive/MyDrive/')

## Open country codes for the Americas and the Caribbean

# output_path = '/content/drive/MyDrive/2023_ODS_OPS/Data/'
output_path = '/content/drive/MyDrive/' # Jimmy
file_name   = 'CountryCodes_Americas.xlsx'
file_path = os.path.join(output_path, file_name)

countries = pd.read_excel(file_path)
#df = df.drop_duplicates(subset='Country Name')
countries.head()
#len(countries["Country Code"].unique())

# countries_code = countries['Country Code'].tolist()
countries_code = countries['Code'].tolist() # Jimmy

## **Retrieving WHO indicator data**

Download an indicator by specifying the indicator code. This will return all associated data for that specific indicator.

Available dimenstions http://ghoapi.azureedge.net/api/WHOSIS_000001

def get_indicator_data(indicator_code):
    url = f"https://ghoapi.azureedge.net/api/{indicator_code}"
    try:
        response = requests.get(url)
        response.raise_for_status()

        data = response.json()["value"]

        # Create empty lists to store the data
        IDs = []
        IndicatorCodes = []
        Countries = []
        years  = []
        values = []
        sex    = []
        Date1  = []
        Date2  = []


        # Extract the data and populate the lists
        for item in data:
            Id = item["Id"]
            IndicatorCode = item["IndicatorCode"]
            year          = item["TimeDimensionValue"]
            value         = item["NumericValue"]
            Country       = item["SpatialDim"]
            Sexo          = item["Dim1"]
            Date11        = item["TimeDimensionBegin"]
            Date22        = item["TimeDimensionEnd"]

            IDs.append(Id)
            IndicatorCodes.append(IndicatorCode)
            Countries.append(Country)
            years.append(year)
            values.append(value)
            sex.append(Sexo)
            Date1.append(Date11)
            Date2.append(Date22)

        # Create the dataframe
        df = pd.DataFrame({"Id": IDs,"SEX": sex, "IndicatorCode": IndicatorCodes, "Year": years, "Date1": Date1,"Date2":Date2,"Value": values, "Country": Countries})

        return df

    except requests.exceptions.RequestException as e:
        print("Error occurred:", e)



## **Crude suicide rates (per 100 000 population) (SDG 3.4.2)**

The age-standardized mortality rate is a weighted average of the age-specific mortality rates per 100 000 persons, where the weights are the proportions of persons in the corresponding age groups of the WHO standard population.
</br>
</br>

**Other related indicator**</br>
Age-standardized suicide rates (per 100 000 population) (MH_12): </br>
</br>
The age-standardized mortality rate is a weighted average of the age-specific mortality rates per 100 000 persons, where the weights are the proportions of persons in the corresponding age groups of the WHO standard population.

#https://www.who.int/data/gho/data/themes/mental-health/suicide-rates
suicide_index         = "SDGSUICIDE"

# Call the function with the URL
df = get_indicator_data(suicide_index)

## Filter countries in the Americas and the Caribbean
suicide_info = df[df['Country'].isin(countries_code)]

##Filtrar solo informacion para ambos sexos
suicide_info = suicide_info[suicide_info['SEX']=='BTSX']

suicide_info1 = suicide_info[['IndicatorCode','Year','Country','Value']]
# Adding the 'IndicatorName' column with the value "Suicide"
suicide_info1['IndicatorName'] = "Crude Suicide Rate"
suicide_info1.head(5)


suicide_info.to_excel('/content/drive/MyDrive/2023_ODS_OPS/SuicideRate.xlsx')

## **Selected WHO Indicators**


#List of selected indicators
path_selected_indicators ='/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/3.List_of_Selected_Suicide_Indicators.xlsx'
SheetName = '3.Suicide_Selected_Indicators'

selected_indicators = pd.read_excel(path_selected_indicators,sheet_name=SheetName )
selected_index_code = selected_indicators['IndicatorCode']


#Indicator 0 is from WHO

index_code = 'FINPROTECTION_IMP_NPRELPL_POP' #selected_index_code[2]
df         = get_indicator_data(index_code)

## Filter countries in the Americas and the Caribbean
Americas_info = df[df['Country'].isin(countries_code)]
Americas_info.head(5)

##Filtrar solo informacion para ambos sexos
Americas_info = Americas_info[Americas_info['SEX']=='TOTL']
Americas_info1= Americas_info[['IndicatorCode','Year','Country','Value']]

# Adding the 'IndicatorName' column with the value "Suicide"
Americas_info1['IndicatorName'] = "Pop below a relative poverty line by household health expeditures"
Americas_info1.head(5)

# Get the unique values of the 'Year' column
unique_years = Americas_info['Year'].unique()

# Sort the unique years in ascending order
sorted_unique_years = sorted(unique_years)
print(sorted_unique_years)

## Identify the missing countries

def find_unique_values(list1, list2):
  set1 = set(list1)
  set2 = set(list2)
  unique_values = list(set1.symmetric_difference(set2))
  return unique_values

unique_countries = Americas_info['Country'].unique()

result = find_unique_values(unique_countries, countries_code)
print('Total countries with data',len(unique_countries))

countries_name = countries[countries['Country Code'].isin(result)]
print(countries_name)


## PAHO Indicators

The excel with the whole information is dowloaded from the PAHO
https://opendata.paho.org/en/core-indicators/download-dataset</br>
</br>


output_path = '/content/drive/MyDrive/2023_ODS_OPS/Data/'
file_name   = 'PAHO-Core-indicators-2023.csv'

file_path = os.path.join(output_path, file_name)

#PAHO_DATA = pd.read_csv(file_path, error_bad_lines=False)
PAHO_DATA = pd.read_csv(file_path)
PAHO_DATA.head(5)

filtered_df = PAHO_DATA[PAHO_DATA['SpatialDim'].isin(countries_code)]
final_df    = filtered_df[['paho_indicator_id',	'indicator_name', 'SpatialDim',	'SpatialDim_en', 'TimeDim','Numeric_value']]

## Explore particular indicators

#List of selected indicators
path_selected_indicators ='/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/3.List_of_Selected_Suicide_Indicators.xlsx'
SheetName = '3.Suicide_Selected_Indicators'

selected_indicators = pd.read_excel(path_selected_indicators,sheet_name=SheetName )
paho_selected_index_code = selected_indicators['IndicatorCode']


index_code = 1304 #paho_selected_index_code[281]
PAHO_info  = final_df[ final_df['paho_indicator_id']==index_code]
sorted(PAHO_info['TimeDim'].unique())

### Create a dataframe with all years and missing information

#indicator from 3 to 31 are from PAHO

# Define an empty list to store the results for each indicator
indicator_results = []


for i in range(3,43):
  #Filter particular indicator
  index_code = selected_index_code[i]
  PAHO_info  = final_df[final_df['paho_indicator_id']==index_code]

  #Fing years availables
  unique_years = PAHO_info['TimeDim'].unique()
  sorted_unique_years = sorted(unique_years)
  first_year = sorted_unique_years[0]
  last_year  = sorted_unique_years[-1]
  Years = f"{first_year}-{last_year}"

  if len(unique_years)>1:
    #Find frequency or periods
    differences = [unique_years[i+1] - unique_years[i] for i in range(len(unique_years) - 1)]
    frequency_count = Counter(differences)
    most_common_difference = frequency_count.most_common(1)
    frequency = most_common_difference[0][0]
  else:
    frequency = 0


  ##identify how many countries have data
  def find_unique_values(list1, list2):
    set1 = set(list1)
    set2 = set(list2)
    unique_values = list(set1.symmetric_difference(set2))
    return unique_values

  unique_countries = PAHO_info['SpatialDim'].unique()
  Total_countries = len(unique_countries)

  ##Names of missing Countries
  missing =  find_unique_values(unique_countries, countries_code)
  missing_countries = countries[countries['Country Code'].isin(missing)]
  missing_countries_names = missing_countries['Country Name'].tolist()

  #Identify missing values
  total_empty_rows = PAHO_info['Numeric_value'].isna().sum()

  # Append the results for the current indicator to the list
  indicator_results.append({
        'index': index_code,
        'Missing countries': missing_countries_names,
        'Countries with Data': Total_countries,
        'Periodicidad Years available': frequency,
        'Years available': Years,
        'Missing data': total_empty_rows
    })

# Create a DataFrame from the list of indicator results
df = pd.DataFrame(indicator_results)
df.head(5)

df.to_excel('/content/drive/MyDrive/2023_ODS_OPS/missingValues.xlsx')

## Filter indicators of interest from PAHO Data

paho_final_indicators = paho_selected_index_code[1:41]

PAHO_Values = PAHO_DATA[PAHO_DATA['paho_indicator_id'].isin(paho_final_indicators)]
PAHO_Values1 = PAHO_Values[['paho_indicator_id','TimeDim','SpatialDim','Numeric_value','indicator_name']]

# Define a dictionary for mapping old column names to new column names
new_column_names = {
    'paho_indicator_id': 'IndicatorCode',
    'TimeDim': 'Year',
    'SpatialDim': 'Country',
    'Numeric_value': 'Value',
    'indicator_name': 'IndicatorName'
}

# Use the 'rename' method to change column names on the copied DataFrame
PAHO_Values1.rename(columns=new_column_names, inplace=True)

PAHO_Values1.head(5)

# Concatenate the DataFrames vertically
combined_df = pd.concat([Americas_info1, PAHO_Values1], ignore_index=True)

print(combined_df.shape)
combined_df.head(5)

combined_df.to_excel('/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/4.Values_Suicide_rate_Indicators_PAHO.xlsx')

# The world Bank

There some additional data dowloaded from
https://databank.worldbank.org/source/world-development-indicators/preview/on#

## Children-Adolescent data from WHO Indicators with 'children/adolescent' keys
08/23/2023
Jimmy


# Children-Adolescent data
# 08/23/2023
# Jimmy

index_codes = ['CC_7', 'CC_8', 'AIR_8', 'CC_5',
               'CC_6', 'AIR_4', 'AIR_10', 'AIR_12', 'AIR_16', 'AIR_6',
               'RSUD_10', 'RSUD_340', 'RSUD_350', 'RSUD_890', 'RSUD_880',
               'RSUD_900', 'TOTENV_11', 'TOTENV_8', 'TOTENV_10', 'TOTENV_12',
               'TOTENV_7', 'TOTENV_9', 'AIR_71', 'NCD_CCS_PA_GUIDELINES_5to19',
               'NCD_CCS_PA_GUIDELINES_UNDER5', 'AIR_72', 'AIR_75', 'AIR_73',
               'CHILDVIOL', 'AIR_74']
children_ad_data = []
for index_code in index_codes:
  print(f'Processing {index_code} ...')
  df = get_indicator_data(index_code)
  ## Filter countries in the Americas and the Caribbean
  df = df[df['Country'].isin(countries_code)]
  children_ad_data.append(df)
children_ad_data = pd.concat(children_ad_data, ignore_index=True)
children_ad_data

indicator_descs_df = pd.DataFrame({
    "IndicatorDesc": [
        "Climate change attributable DALYs ('000) in children under 5 years",
        "Climate change  attributable DALYs per 100'000 children under 5 years",
        "Ambient air pollution attributable DALYs  in children under 5 years",
        "Climate change attributable deaths ('000) in children under 5 years",
        "Climate change  attributable deaths per 100'000 children under 5 years",
        "Ambient air pollution attributable deaths in children under 5 years",
        "Ambient air pollution  attributable DALYs per 100'000 children under 5 years",
        "Household air pollution attributable deaths in children under 5 years",
        "Household air pollution attributable DALYs in children under 5 years",
        "Ambient air pollution  attributable deaths per 100'000 children under 5 years",
        "National survey on substance use among children and adolescents",
        "Treatment programmes for children and adolescents with alcohol use disorders",
        "Treatment programmes for children and adolescents with drug use disorders",
        "Treatment programmes for children and adolescents with alcohol use disorders",
        "Epidemiological data collection system for substance use among children and adolescents",
        "Treatment programmes for children and adolescents with drug use disorders",
        "Deaths attributable to the environment in children under 5 years (%)",
        "Total environment  attributable deaths per 100'000 children under 5 years",
        "Total environment  attributable DALYs per 100'000 children under 5 years",
        "DALYs attributable to the environment in children under 5 years (%)",
        "Total environment attributable deaths ('000) in children under 5 years",
        "Total environment attributable DALYs ('000) in children under 5 years",
        "Ambient air pollution attributable DALYs in children under 15 years",
        "Existence of national guidelines for physical activity for children and adolescents aged 5-19",
        "Existence of national guidelines for physical activity for children under 5",
        "Ambient air pollution attributable deaths in children under 15 years",
        "Household air pollution attributable DALYs in children under 15 years",
        "Ambient air pollution attributable YLL in children under 15 years",
        "Indicator 16.2.1: Proportion of children aged 1-17 years who experienced any physical punishment and/or psychological aggression by caregivers in the past month",
        "Household air pollution attributable deaths in children under 15 years"],
    "IndicatorCode": index_codes
    })
indicator_descs_df

children_ad_data = children_ad_data.merge(indicator_descs_df, left_on='IndicatorCode', right_on='IndicatorCode', how='left')
children_ad_data.to_excel('/content/drive/MyDrive/WHO_children_adolescent_data.xlsx')
