==> 1.DataCollection/3.JoinFiles.ipynb <==
<center>

# **Join Files**

</center>

This algorith join all the indicators downloaded from
https://ghdx.healthdata.org/record/global-burden-disease-study-2019-gbd-2019-covariates-1980-2019


import pandas as pd
from prettytable import PrettyTable
import matplotlib.pyplot as plt
from collections import Counter
import requests
import io
import os

#Mount your Google Drive to Colab

from google.colab import drive
drive.mount('/content/drive')

## Health Indicators filtered from WHO database

output_path = '/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/'
file_name   = '2.Values_Health_Coverage_Indicators_WHO.xlsx'
file_path = os.path.join(output_path, file_name)

health_coverage_data = pd.read_excel(file_path)


health_coverage_data_filter = health_coverage_data[['IndicatorCode','SEX','Year','Country','Value']]
health_coverage_data_filter.head(5)

## Step 1: Join all files in folder GBD_1980_2019_Data

folder_path = '/content/drive/MyDrive/2023_ODS_OPS/Data/GBD_1980_2019_Data/'

# Get a list of all Excel files in the folder
excel_files = [file for file in os.listdir(folder_path) if file.endswith('.xlsx')]


#estos archivos #0,1,2,4,11 tienen information por grupo de edad

#file= excel_files[11]
#file_path = os.path.join(folder_path, file)
#df = pd.read_excel(file_path)
#print(df['IndicatorName'].unique())
#df.head(5)

reduce_excel_files = [excel_files[i] for i in [3,5,6,7,8,9,10]]

# Initialize an empty list to store DataFrames
dfs = []

# Read each Excel file and store it in the list
for file in reduce_excel_files:
    file_path = os.path.join(folder_path, file)
    df = pd.read_excel(file_path)
    dfs.append(df)


# Concatenate all DataFrames into a single DataFrame
combined_df = pd.concat(dfs, ignore_index=True)
combined_df.head(5)

### Save indicator names

filter_df = combined_df[['IndicatorName','IndicatorCode']]
# Drop duplicated rows based on the selected columns
filter_df = filter_df.drop_duplicates()
filter_df.head(5)

# Path to the output file where you want to save the combined data
output_file = '/content/drive/MyDrive/2023_ODS_OPS/list_extra_indicators.xlsx'

# Save the combined DataFrame to a new Excel file
filter_df.to_excel(output_file, index=False)


## Filter indicator related with health coverage and join with WHO Base

new_indicators=combined_df[['IndicatorCode',	'sex_label', 'Year',	'Country',	'Value']]
new_indicators.head(5)

# Change the name of the column from 'IndicatorName' to 'SEX'
new_indicators.rename(columns={'sex_label': 'SEX'}, inplace=True)

# Update the values 'female', 'male', and 'Both' to 'FMLE', 'MLE', and 'BOTH' respectively
new_indicators['SEX'] = new_indicators['SEX'].replace({
    'Female': 'FMLE',
    'Male': 'MLE',
    'Both': ''
})

# Filtering the data
new_indicators_filter_year = new_indicators[(new_indicators['Year'] >= 2000) & (new_indicators['Year'] <= 2019)]
new_indicators_filter_year.head(5)


new_indicators_filter_year.to_excel('/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/5.Indicators_from_ghdx_healthdata.xlsx')

list_indicators_to_join_health_data = ['pop_dens_300_500_psqkm_pct', 'pop_dens_500_1000_psqkm_pct', 'pop_dens_over_1000_psqkm_pct', 'pop_dens_150_300_psqkm_pct', 'pop_dens_under_150_psqkm_pct', 'haqi']

indicators_health = new_indicators[new_indicators['IndicatorCode'].isin(list_indicators_to_join_health_data)]
indicators_health.head(5)

indicators_health['Year'].unique()

final_healht_data = pd.concat([health_coverage_data_filter,indicators_health])
final_healht_data['SEX'].unique()


# Filtering the data
final_healht_data_year_filter = final_healht_data[(final_healht_data['Year'] >= 2000) & (final_healht_data['Year'] <= 2019)]


final_healht_data_year_filter.to_excel('/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/2.Values_Health_Coverage_Indicators_WHO_gdhx_healthdata.xlsx')

final_healht_data_year_filter.to_csv('/content/drive/MyDrive/2023_ODS_OPS/OutputFiles/2.Values_Health_Coverage_Indicators_WHO_gdhx_healthdata.csv')
